Abtracted out Transform Builder
- Abstracted out the transform builder so that it now returns the correct transforms depending on the type that's being passed in.
- Values were being passed through multiple levels which seemed a little redundant. Abstracted out and hopefully simplified it.
- Renamed some variables so that they make more sense in the context they're being used.
- Removed the `PlaceHolderText` because as far as i could see, it was always `output`.
- Removed `FILTER_MAP` stuff, to a `pipeAnswerCheck`. I think simplifies the code on top of that makes it a bit easier to read.
